removed email form input from login page and added it to signup page.
