+++ client/src/pages/Signup.jsx
@@ -1,20 +1,26 @@
-// see SignupForm.js for comments
 import React, { useState } from "react";
 import { Form, Button, Alert } from "react-bootstrap";
+// import hooks for mutations and our mutations
 import { useMutation } from "@apollo/react-hooks";
-import { LOGIN_USER } from "../utils/mutations";
+import { ADD_USER } from "../utils/mutations";
 import Auth from "../utils/auth";
 
-const LoginForm = () => {
-  const [userFormData, setUserFormData] = useState({ email: "", password: "" });
+const SignupForm = () => {
+  // set initial form state
+  const [userFormData, setUserFormData] = useState({
+    username: "",
+    email: "",
+    password: "",
+  });
+  // set state for form validation
   const [validated] = useState(false);
-  const [showAlert, setShowAlert] = useState(false);
-
   // using the apollo hook  useMutation pass the
   // ADD_USER mutation in order to talk to graphql
   // addUser will hold the output and error the error
 
-  const [login, { error }] = useMutation(LOGIN_USER);
+  const [addUser, { error }] = useMutation(ADD_USER);
+  // set state for alert
+  const [showAlert, setShowAlert] = useState(false);
 
   const handleInputChange = (event) => {
     const { name, value } = event.target;
@@ -31,16 +37,20 @@ const LoginForm = () => {
       event.stopPropagation();
     }
 
+    if (error) {
+      throw new Error("something went wrong!");
+    }
+
+    // use try/catch instead of promises to handle errors
     try {
-      const { data } = await login({
+      // execute addUser mutation and pass in variable data from form
+      const { data } = await addUser({
         variables: { ...userFormData },
       });
-
-      // Store the token in local storage
-      Auth.login(data.login.token);
+      Auth.login(data.addUser.token);
       console.log(data);
-    } catch (e) {
-      console.error(e);
+    } catch (err) {
+      console.error(err);
       setShowAlert(true);
     }
 
@@ -53,20 +63,38 @@ const LoginForm = () => {
 
   return (
     <>
+      {/* This is needed for the validation functionality above */}
       <Form noValidate validated={validated} onSubmit={handleFormSubmit}>
+        {/* show alert if server response is bad */}
         <Alert
           dismissible
           onClose={() => setShowAlert(false)}
           show={showAlert}
           variant="danger"
         >
-          Something went wrong with your login credentials!
+          Something went wrong with your signup!
         </Alert>
+
+        <Form.Group>
+          <Form.Label htmlFor="username">Username</Form.Label>
+          <Form.Control
+            type="text"
+            placeholder="Your username"
+            name="username"
+            onChange={handleInputChange}
+            value={userFormData.username}
+            required
+          />
+          <Form.Control.Feedback type="invalid">
+            Username is required!
+          </Form.Control.Feedback>
+        </Form.Group>
+
         <Form.Group>
           <Form.Label htmlFor="email">Email</Form.Label>
           <Form.Control
-            type="text"
-            placeholder="Your email"
+            type="email"
+            placeholder="Your email address"
             name="email"
             onChange={handleInputChange}
             value={userFormData.email}
@@ -92,17 +120,21 @@ const LoginForm = () => {
           </Form.Control.Feedback>
         </Form.Group>
         <Button
-          disabled={!(userFormData.email && userFormData.password)}
+          disabled={
+            !(
+              userFormData.username &&
+              userFormData.email &&
+              userFormData.password
+            )
+          }
           type="submit"
           variant="success"
         >
           Submit
         </Button>
-        {error && <div>Login failed</div>}
       </Form>
     </>
   );
 };
 
-export default LoginForm;
-
+export default SignupForm;
