+++ package.json
@@ -14,6 +14,9 @@
     "license": "ISC",
     "devDependencies": {
       "concurrently": "^5.3.0"
-    }
+    },
+    "dependencies": {
+      "graphql": "^16.8.1"
   }
+}
