+++ server/routes/api/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const userRoutes = require('./user-routes');
+
+router.use('/users', userRoutes);
+
+module.exports = router;
